# Optimize time cost of on_raft_gc_log_tick