--- a/src/raftstore/store/worker/mod.rs
+++ b/src/raftstore/store/worker/mod.rs
@@ -61,7 +61,9 @@ pub use self::apply::{
 pub use self::cleanup_sst::{Runner as CleanupSSTRunner, Task as CleanupSSTTask};
 pub use self::compact::{Runner as CompactRunner, Task as CompactTask};
 pub use self::consistency_check::{Runner as ConsistencyCheckRunner, Task as ConsistencyCheckTask};
-pub use self::raftlog_gc::{Runner as RaftlogGcRunner, Task as RaftlogGcTask};
+pub use self::raftlog_gc::{
+    BaseRaftLogGcTask, EntryGcKey, EntryGcTask, Runner as RaftlogGcRunner, Task as RaftlogGcTask,
+};
 pub use self::read::{LocalReader, Progress as ReadProgress, Task as ReadTask};
 pub use self::region::{
     Runner as RegionRunner, Task as RegionTask, PENDING_APPLY_CHECK_INTERVAL,