<!--
Thank you for contributing to TiKV! Please read TiKV's [CONTRIBUTING](https://github.com/tikv/tikv/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

## What have you changed? (mandatory)

Please explain **IN DETAIL** what the changes are in this PR and why they are needed:
- Optimize time cost of on_raft_gc_log_tick by using another thread to reclaim memory of Entry.
- 0.99 histogram_quantile of `tikv_raftstore_event_duration_bucket` reduces from maximum 64ms to 25.5ms.
- AVG of `tikv_raftstore_event_duration` reduces from maximum 21.8ms to 9ms.
- May bring more memory fragmentation.

## What are the type of the changes? (mandatory)

The currently defined types are listed below, please pick one of the types for this PR by removing the others:
- Improvement

## How has this PR been tested? (mandatory)
- stress test
- detail in pingcap-jira-TIKV-1611

## Does this PR affect documentation (docs/docs-cn) update? (mandatory)

## Does this PR affect tidb-ansible update? (mandatory)

## Refer to a related PR or issue link (optional)
- Based on [PR 3657](https://github.com/tikv/tikv/pull/3657)

## Benchmark result if necessary (optional)

## Add a few positive/negative examples (optional)