@zhangjinpeng1987 2018-10-14
I wonder if it can reduce max latency under heavy write load?