@solotzg 2018-10-14
@zhangjinpeng1987 
In my [stress test](https://internal.pingcap.net/jira/browse/TIKV-1611?focusedCommentId=22175&page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel#comment-22175), improvement about qps and latency is limited. Maybe  because I used the default setting of tikv. Further tests may be needed.