@huachaohuang 2018-10-15
IMO, this is not a worthy solution, it introduces complexity but only reduces the latency of `RaftLogGc` to 25.5ms (which is still not acceptable). I think we should move these timeout tasks out of the critical path in the long term. For now, can we do some tiny changes to ease the problem? Like using some other data structure?