@siddontang 2018-10-16
> The time cost of system memory reclaiming grows when the length of data and context grows.

Can you add a benchmark here? 

Seem that we can also meet the same problem elsewhere, sadly.