@huachaohuang 2018-11-20
I think we can keep it open until we figure out if we have better solutions.